Database model was changed to version 2 Domain was adjusted for new database model Course name now is course level attribute Endpoints have been implemented to obtain evaluations and workshops of the students in a subject in a particular course Endpoints have been implemented to obtain the evaluations and workshops carried out in a subject in a given course

+++ src/main/java/com/ufro/culmingapp/school/domain/SchoolRepository.java
@@ -2,6 +2,6 @@ package com.ufro.culmingapp.school.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SchoolRepository extends JpaRepository<School, Integer> {
+public interface SchoolRepository extends JpaRepository<School, Long> {
 
 }
